+++ example_project_app/plugins.py
@@ -1,10 +1,11 @@
+from django.conf import settings
 from django.urls import reverse
 
 # Projectroles dependency
 from projectroles.models import SODAR_CONSTANTS
 from projectroles.plugins import ProjectAppPluginPoint
 
-from .urls import urlpatterns
+from example_project_app.urls import urlpatterns
 
 
 class ProjectAppPlugin(ProjectAppPluginPoint):
@@ -232,3 +233,31 @@ class ProjectAppPlugin(ProjectAppPluginPoint):
                 'url': reverse('home'),
             },
         }
+
+    def get_app_alerts(self, user):
+        if not settings.EXAMPLE_PROJECT_APP_ALERTS:
+            return None
+        return [
+            {
+                'message': 'This is an example app alert with HTML. '
+                '<a href="{}">This is a link</a>.'.format(reverse('home')),
+                'url': reverse('home'),
+                'level': 'INFO',
+            },
+            {
+                'message': 'Another example, this time at the SUCCESS level.',
+                'url': reverse('home'),
+                'level': 'SUCCESS',
+            },
+            {
+                'message': 'An example WARNING alert without an URL.',
+                'url': None,
+                'level': 'WARNING',
+            },
+            {
+                'message': 'Finally, a DANGER alert. Hi there '
+                '<a href="mailto:{}">{}</a>!'.format(user.email, user.username),
+                'url': reverse('home'),
+                'level': 'DANGER',
+            },
+        ]