+++ CHANGELOG.rst
@@ -13,6 +13,8 @@ Unreleased
 Added
 -----
 
+- **General**
+    - Example app alerts in ``example_project_app`` (#642)
 - **Adminalerts**
     - ``get_statistics()`` implementation
 - **Projectroles**
@@ -27,6 +29,9 @@ Added
     - ``get_statistics()`` in ``SiteAppPluginPoint``
     - ``info_settings`` in app plugins (#671)
     - ``plugin_type`` argument in ``get_app_plugin()`` (#309)
+    - Support for site-wide app alerts (#642)
+    - Site app plugin ``AppAlertSiteAppPlugin`` for app alerts (#642)
+    - ``get_app_alerts()`` in project and site app plugin points (#642)
 - **Siteinfo**
     - Display selected Django settings in UI (#671)
 - **Timeline**